Moved the pass to hls/opt
