+++ jlm/hls/Makefile.sub
@@ -22,7 +22,6 @@ libhls_SOURCES = \
     jlm/hls/backend/rvsdg2rhls/distribute-constants.cpp \
     jlm/hls/backend/rvsdg2rhls/GammaConversion.cpp \
     jlm/hls/backend/rvsdg2rhls/instrument-ref.cpp \
-    jlm/hls/backend/rvsdg2rhls/InvariantLambdaMemoryStateRemoval.cpp \
     jlm/hls/backend/rvsdg2rhls/mem-conv.cpp \
     jlm/hls/backend/rvsdg2rhls/mem-queue.cpp \
     jlm/hls/backend/rvsdg2rhls/mem-sep.cpp \
@@ -38,6 +37,7 @@ libhls_SOURCES = \
     jlm/hls/ir/hls.cpp \
     \
     jlm/hls/opt/cne.cpp \
+    jlm/hls/opt/InvariantLambdaMemoryStateRemoval.cpp \
     \
     jlm/hls/util/view.cpp \
 
@@ -62,7 +62,6 @@ libhls_HEADERS = \
 	jlm/hls/backend/rvsdg2rhls/distribute-constants.hpp \
 	jlm/hls/backend/rvsdg2rhls/GammaConversion.hpp \
 	jlm/hls/backend/rvsdg2rhls/instrument-ref.hpp \
-	jlm/hls/backend/rvsdg2rhls/InvariantLambdaMemoryStateRemoval.hpp \
 	jlm/hls/backend/rvsdg2rhls/mem-conv.hpp \
 	jlm/hls/backend/rvsdg2rhls/mem-queue.hpp \
 	jlm/hls/backend/rvsdg2rhls/mem-sep.hpp \
@@ -78,12 +77,12 @@ libhls_HEADERS = \
 	jlm/hls/ir/hls.hpp \
 	\
 	jlm/hls/opt/cne.hpp \
+	jlm/hls/opt/InvariantLambdaMemoryStateRemoval.hpp \
 	\
 	jlm/hls/util/view.hpp \
 
 libhls_TESTS += \
 	tests/jlm/hls/backend/rvsdg2rhls/DeadNodeEliminationTests \
-	tests/jlm/hls/backend/rvsdg2rhls/InvariantLambdaMemoryStateRemovalTests \
 	tests/jlm/hls/backend/rvsdg2rhls/MemoryConverterTests \
 	tests/jlm/hls/backend/rvsdg2rhls/MemoryQueueTests \
 	tests/jlm/hls/backend/rvsdg2rhls/TestFork \
@@ -92,6 +91,8 @@ libhls_TESTS += \
 	tests/jlm/hls/backend/rvsdg2rhls/UnusedStateRemovalTests \
 	tests/jlm/hls/backend/rvsdg2rhls/test-loop-passthrough \
 	tests/jlm/hls/util/ViewTests \
+	\
+	tests/jlm/hls/opt/InvariantLambdaMemoryStateRemovalTests \
 
 libhls_TEST_LIBS += \
 	libjlmtest \
