+++ jlm/hls/opt/InvariantLambdaMemoryStateRemoval.hpp
